requirements.txt
pandas
requests
sqlalchemy
python-dotenv
pyodbc
